Renamed projectview to project for clarity
