+++ src/components/Project/ProjectBar.vue
